+++ CHANGELOG.md
@@ -69,6 +69,7 @@ The following changes have been completed, and will be included in the next tagg
 - added `hues()` inspector function
 - added `isBlackHue()` inspector function
 - added `isBlueHue()` inspector function
+- added `isCyanHue()` inspector function
 - added `isGrayHue()` inspector function
 - added `isDark()` inspector function
 - added `isDull()` inspector function